+++ src/Components/Feed/Comment.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import elapsedTime from './elapsedTime';
 import { useNavigate } from 'react-router-dom';
 import { APIDefaultImage, profileImg } from './COMMON';
-import useFetchComment from './useFetchComment';
+import useFetchComment from '../../Hooks/useFetchComment';
 const Comment = ({ feedList, commentList, setCommentList, isFetchData, setIsFetchData }) => {
   // useEffect(() => {
   //   setIsFetchData(false); // 컴포넌트가 리렌더링될 때마다 데이터를 다시 불러오기 위해 isFetchData 상태를 false로 설정
